Move PDFPage to PDFPage.swift
